API reference

GeometricAlgebra.GA20

GeometricAlgebra.GA24

GeometricAlgebra.GA30

GeometricAlgebra.GA3232

GeometricAlgebra.GA33

GeometricAlgebra.GA40

GeometricAlgebra.GA44

GeometricAlgebra.PGA

GeometricAlgebra.Quaternions

GeometricAlgebra.STA

GeometricAlgebra.project

Core functions

GeometricAlgebra.project — Function

project(multivector::T, n::Integer) -> T

Project a given multivector of type T so that only the component of grade n is kept.

The function will always return a multivector of the same type as that passed in. If n is such that multivector does not contain this grade (e.g. because n is odd, and the multivector is an even type), then we will return a zero multivector.

Modules

There is one submodule for each algebra we support. These are listed below:

GeometricAlgebra.GA20 — Module

GA20

Module representing the GA(2, 0) geometric algebra. Even and odd elements are stored as complex numbers.

GeometricAlgebra.GA24 — Module

GA24

Module representing the GA(2,4) geometric algebra.

This is the conformal algebra for spacetime. Even and odd elements are stored as complex Static Arrays.

GeometricAlgebra.GA30 — Module

GA30

Module representing the GA(3, 0) geometric algebra. Underlying rep is quaternions, but not used explicitly here to keep this code self-contained.

GeometricAlgebra.GA33 — Module

GA33

Module for GA(3,3). Base representation is a pair of 4x4 static arrays.

GeometricAlgebra.GA40 — Module

GA40

Code for GA(4,0). Even and odd elements are stored as quaternion pairs.

GeometricAlgebra.GA44 — Module

GA44

Uses a list-based approach to multivector storage. Not as fast as smaller algebras, but useful for checks.

GeometricAlgebra.GA3232 — Module

GA3232

Module representing the GA(32,32) geometric algebra.

Not optimised in any way, but useful for some experimentation. Use with caution. GA(32,32) is very BIG!

GeometricAlgebra.PGA — Module

PGA

Code for GA(3,0,1). The Euclidean group algebra. C Core representation is a quaternion pair.

GeometricAlgebra.STA — Module

STA

Code for GA(1,3). Even and odd elements are stored as complex structs. 2x2 matrix products are unwrapped. Gamma and sigma used for pretty typing.

Quaternions

Note

This submodule is considered an implementation detail for GeometricAlgebra.

GeometricAlgebra.Quaternions — Module

Quaternions

Quaternion code. This is called by some of the later GA implementations.

The core mirrors much of the GA code structure. For completeness we have defined a division operation for quaternions as they are a division algebra.